echo.c - adding an empty space for the first fprintf and removing the if statement
Adding an empty space for the first fprintf and removing the if statement. The purpose is to eliminate argc-1 executions of the if statement and argc-1 calls of fprintf. The result of the echo will be the same as before but the produced machine code will have less instructions and less function calls for more performance.